gateway/blueprints-with-process-api/requirements_with_version.txt
requests==2.22.0
prometheus_client==0.7.1
flask==1.1.2
jinja2==2.11.0
gunicorn==20.0.4
gevent==20.5.0
Pillow==7.2.0
opencv-python==4.2.0.34
tensorflow-serving-api==2.1.0